// web-client/src/types.d.ts
export interface ArtistType {
  id: number;
  name: string;
  biography: string | null;
  countryOfOrigin: string | null;
  debutDate: string | null;
  imageUrl: string | null;
  type: string | null;
}

export interface ArtistContextType {
  artistsList: ArtistType[];
  addArtist: (newArtist: Omit<ArtistType, "id">) => void;
  deleteArtist: (id: number) => void;
  getArtistById: (id: number) => ArtistType | null;
  updateArtist: (id: number, updatedArtist: Partial<ArtistType>) => void;
  searchArtists: (params: Record<string, string>) => void;
}

export interface RecordLabelType {
  id: number;
  name: string;
  country: string | null;
  foundationDate: string | null;
  website: string | null;
  logoUrl: string | null;
}
// ...
export interface AlbumType {
  id: number;
  artistId: Partial<ArtistType>;
  recordLabelId: Partial<RecordLabelType>;
  title: string;
  releaseDate: string;
  coverUrl: string | null;
  type: string | null;
  totalDuration: number | null;
  description: string | null;
}

export interface AlbumContextType {
  albumsList: AlbumType[];
  addAlbum: (newAlbum: Omit<AlbumType, "id" | "totalDuration">) => void;
  deleteAlbum: (id: number) => void;
  getAlbumById: (id: number) => AlbumType | null;
  updateAlbum: (id: number, updatedAlbum: Partial<AlbumType>) => void;
  searchAlbums: (params: Record<string, string>) => void;
}

export interface GenreType {
  id: number;
  name: string;
  description: string | null;
}

export interface GenreContextType {
  genresList: GenreType[];
  addGenre: (newGenre: Omit<GenreType, "id">) => void;
  deleteGenre: (id: number) => void;
  getGenreById: (id: number) => GenreType | null;
  updateGenre: (id: number, updatedGenre: Partial<GenreType>) => void;
  searchGenres: (params: Record<string, string>) => void;
}

export interface SongType {
  id: number;
  albumId: Partial<AlbumType> | null;
  artistId: Partial<ArtistType>;
  genreId: Partial<GenreType> | null;
  title: string;
  composer: string | null;
  duration: number;
  lyrics: string | null;
  releaseDate: string | null;
  fileUrl: string | null;
  imageUrl: string | null;
}